pyproject.toml
[project]
name = "datatalent-end-to-end"
version = "0.1.0"
description = "Pipeline DataTalent — ingestion GCP (scripts/) et transformations dbt"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "altair>=5.5.0",
    "db-dtypes>=1.4.0",
    "google-cloud-bigquery>=3.40.1",
    "google-cloud-storage>=3.9.0",
    "loguru>=0.7.3",
    "pandas>=3.0.1",
    "pyarrow>=23.0.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "streamlit>=1.40.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]
